+++ FE/src/apis/calendar.ts
@@ -1,7 +1,26 @@
+import { SimpleCalendarDto } from "@/apis/dtos/calendar.dto";
 import { https } from "@/apis/instance";
 import API from "@/constants/API";
 import { generateCurrentMonthMockData } from "@/mocks/calendarMockData";
 import { DateFilter, NewCalendar } from "@/types/calendar";
+
+export const getCalendarEventsOnWeek = async (): Promise<
+  SimpleCalendarDto[]
+> => {
+  const { data } = await https({
+    method: "GET",
+    url: API.CALENDAR.WEEKLY({
+      year: new Date().getFullYear(),
+      month: new Date().getMonth() + 1,
+      date: new Date().getDate(),
+      duration: 14,
+    }),
+  });
+
+  return data.data.calendars.map(
+    (item: SimpleCalendarDto) => new SimpleCalendarDto({ ...item }),
+  );
+};
 
 export function postCalender(newCalendar: NewCalendar) {
   return https({
